src/main/java/com/example/restblogapp/service/UserService.java
package com.example.restblogapp.service;

import com.example.restblogapp.entities.User;
import com.example.restblogapp.repositories.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.context.annotation.Bean;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class UserService {

    @Autowired
    private UserRepository userRepository;

    @Bean
    public PasswordEncoder passwordEncoder(){
        return new BCryptPasswordEncoder();
    }

    public void save(User user){
        user.setPassword(passwordEncoder().encode(user.getPassword()));
        userRepository.save(user);
    }

    public User getUser(String username){
        return userRepository.findByUsername(username);
    }

    public List<User> getAllUsers() {
        return userRepository.findAll();
    }

    public User login(String username, String password) {
        var userOp = userRepository.findByUsernameAndPassword(username, password);
        if (userOp.isPresent()) {
            return userOp.get();
        } else {
            throw new IllegalArgumentException("Invalid Login details");
        }
    }
}
